refactor according to PR comments
