+++ graphqlapi/src/main/java/io/stargate/graphql/web/resources/schemafirst/SchemaFirstResourceBase.java
@@ -1,0 +1,20 @@
+/*
+ * Copyright The Stargate Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.stargate.graphql.web.resources.schemafirst;
+
+import io.stargate.graphql.web.resources.GraphqlResourceBase;
+
+public class SchemaFirstResourceBase extends GraphqlResourceBase {}
